+++ helpers/request.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"strings"
 )
 
 func GetJsonFromRequestBody(r *http.Request, v interface{}) error {
@@ -46,4 +47,17 @@ func WriteErrorJson(w http.ResponseWriter, r *http.Request, e error) {
 	}
 	resp.Error.Message = e.Error()
 	render.JSON(w, r, resp)
+}
+
+func GetJwtToken(r *http.Request) (string, error) {
+	header := r.Header.Get("Authorization")
+	if header == "" {
+		return "", errors.New("couldn't find authorization header")
+	}
+
+	if !strings.HasPrefix(strings.ToLower(header), "bearer") {
+		return "", errors.New("authorization header didn't start with 'bearer'")
+	}
+
+	return header[7:], nil
 }
